Create Images folder and separate widget to other folder
