+++ quiz_quest/lib/app/features/welcome_page/welcome_page.dart
@@ -1,0 +1,44 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter/material.dart';
+
+
+
+class WelcomePage extends StatelessWidget {
+  const WelcomePage({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: StreamBuilder<QuerySnapshot<Map<String, dynamic>>>(
+          stream: FirebaseFirestore.instance.collection('example').snapshots(),
+          builder: (context, snapshot) {
+            if (snapshot.hasError) {
+              return const Center(
+                child: Text('Something went wrong'),
+              );
+            }
+
+            if (snapshot.connectionState == ConnectionState.waiting) {
+              return const Center(
+                child: Text('Loading'),
+              );
+            }
+
+            final documents = snapshot.data!.docs;
+
+            return Center(
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  for (final doc in documents) ...[
+                    Text(doc['example']),
+                  ],
+                ],
+              ),
+            );
+          }),
+    );
+  }
+}
